loki2nats: A working version with fastwebsockets (async)
How to run:

    APP <hostname> '{host!="", tenant="abcdef"}'

Tests ran against a single (busy) Loki tenant, yielded:

    code    |  rtime    |  utime | stime | handled | dropped | drop% |  total
    --------+-----------+--------+-------+---------+---------+-------+-------
    python  | 01:00.367 | 36.391 | 0.216 |   85292 |   31610 |   27% | 116902
    rust    | 01:00.011 |  9.645 | 0.663 |   91948 |   27915 |   23% | 119863
    rust-rl | 01:00.184 |  2.979 | 0.441 |   61799 |    4823 |    7% |  66622
    --------+-----------+--------+-------+---------+---------+-------+-------

Conclusion, neither is good enough.

(rust-rl is release build. It's better, but got a slightly advantageous
test because the total messages seemed lower then. 7% drop is still too
much.)

Also, the async rust version (a) uses threads for no good reason at the
moment, and (b) it takes a couple of seconds for it to realise there is
data. That is flawed.