loki2nats: A working version with tungstenite (sync)
How to run:

    APP <hostname> '{host!="", tenant="abcdef"}'

Tests ran against a single (busy) Loki tenant, yielded:

    code    |  rtime    |  utime | stime | handled | dropped | drop% |  total
    --------+-----------+--------+-------+---------+---------+-------+-------
    python  | 01:00.367 | 36.391 | 0.216 |   85292 |   31610 |   27% | 116902
    asyncdb | 01:00.011 |  9.645 | 0.663 |   91948 |   27915 |   23% | 119863
    asyncrl | 01:00.184 |  2.979 | 0.441 |   61799 |    4823 |    7% |  66622
    syncdb  | 00:59.952 |  8.787 | 0.478 |   72408 |   12075 |   14% |  84483
    syncrl  | 01:00.256 |  4.008 | 0.482 |   92672 |   46426 |   33% | 139098
    --------+-----------+--------+-------+---------+---------+-------+-------

The multithreading -- which we didn't utilize -- is gone, so that saves
us a few cycles. However it looks like this test with syncrl got a bunch
more messages and hence a bunch more drops. 33% drops is not really
nice.

It looks like we'd have to filter more to drop less, if this is even
possible at all. Maybe doing the websocket /tail is not guaranteed to give us
everything if the pipe is too small. In that case we might need to look
into fetching things on a chunk-by-chunk basis ("give me all of the last
minute").

Interestingly, we _still_ have the startup problem that the Python
version does not have. Something might be stuck at the TLS<->TCP layer.

Also noteworthy: the tungstenite people did claim to be 30% _slower_
than fastwebsockets. On the other hand, we handled about 50% more
messages (+ more drop count handling), so the utime should be considered
a winner here.